Move `FullDef` to its own module
