+++ frontend/exporter/src/types/new/mod.rs
@@ -1,11 +1,13 @@
 //! This module contains type definitions that have no equivalent in
 //! Rustc.
 
+mod full_def;
 mod impl_infos;
 mod item_attributes;
 mod predicate_id;
 mod typed_constant_kind;
 
+pub use full_def::*;
 pub use impl_infos::*;
 pub use item_attributes::*;
 pub use predicate_id::*;
